requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2
django-filter==23.1
django-stubs==1.16.0
django-stubs-ext==0.8.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
flower==1.2.0
humanize==4.6.0
iniconfig==2.0.0
kombu==5.2.4
lxml==4.9.2
mypy==1.1.1
mypy-extensions==1.0.0
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
PyJWT==1.7.1
pytest==7.3.1
pytest-django==4.5.2
pytz==2023.3
redis==4.5.4
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tornado==6.2
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.9
typing_extensions==4.5.0
vine==5.0.0
wcwidth==0.2.6
